Kill `GameData`
