+++ src/main/java/net/neoforged/neoforge/common/data/internal/NeoForgeRegistryOrderReportProvider.java
@@ -12,7 +12,7 @@ import java.util.concurrent.CompletableFuture;
 import net.minecraft.data.CachedOutput;
 import net.minecraft.data.DataProvider;
 import net.minecraft.data.PackOutput;
-import net.neoforged.neoforge.registries.GameData;
+import net.neoforged.neoforge.registries.RegistryManager;
 
 public final class NeoForgeRegistryOrderReportProvider implements DataProvider {
     private final PackOutput output;
@@ -26,7 +26,7 @@ public final class NeoForgeRegistryOrderReportProvider implements DataProvider {
         JsonObject json = new JsonObject();
 
         JsonArray array = new JsonArray();
-        GameData.getRegistrationOrder().forEach(name -> array.add(name.toString()));
+        RegistryManager.getRegistrationOrder().forEach(name -> array.add(name.toString()));
         json.add("order", array);
 
         Path path = this.output.getOutputFolder(PackOutput.Target.REPORTS).resolve("registry_order.json");
