+++ src/main/java/net/neoforged/neoforge/registries/RegistryManager.java
@@ -5,6 +5,7 @@
 
 package net.neoforged.neoforge.registries;
 
+import com.google.common.collect.ImmutableMap;
 import com.mojang.logging.LogUtils;
 import io.netty.util.AttributeKey;
 import java.util.ArrayList;
@@ -13,18 +14,31 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.IdentityHashMap;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
+import net.minecraft.core.Holder;
+import net.minecraft.core.IdMapper;
 import net.minecraft.core.MappedRegistry;
 import net.minecraft.core.RegistrationInfo;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.SpawnPlacements;
+import net.minecraft.world.entity.ai.village.poi.PoiType;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
 import net.neoforged.fml.ModLoader;
+import net.neoforged.fml.loading.progress.StartupNotificationManager;
+import net.neoforged.neoforge.common.CommonHooks;
+import net.neoforged.neoforge.common.CreativeModeTabRegistry;
+import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.neoforge.network.configuration.RegistryDataMapNegotiation;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 import net.neoforged.neoforge.network.payload.FrozenRegistryPayload;
@@ -46,6 +60,18 @@ public class RegistryManager {
     private static Map<ResourceLocation, RegistrySnapshot> vanillaSnapshot = null;
     private static Map<ResourceLocation, RegistrySnapshot> frozenSnapshot = null;
     private static Map<ResourceKey<Registry<?>>, Map<ResourceLocation, DataMapType<?, ?>>> dataMaps = Map.of();
+
+    public static Map<Block, Item> getBlockItemMap() {
+        return NeoForgeRegistryCallbacks.ItemCallbacks.BLOCK_TO_ITEM_MAP;
+    }
+
+    public static IdMapper<BlockState> getBlockStateIdMap() {
+        return NeoForgeRegistryCallbacks.BlockCallbacks.BLOCKSTATE_TO_ID_MAP;
+    }
+
+    public static Map<BlockState, Holder<PoiType>> getBlockStatePointOfInterestTypeMap() {
+        return NeoForgeRegistryCallbacks.PoiTypeCallbacks.BLOCKSTATE_TO_POI_TYPE_MAP;
+    }
 
     /**
      * Called by {@link RegistryBuilder} to make sure that modders don't forget to register their registries.
@@ -104,11 +130,38 @@ public class RegistryManager {
         dataMaps = Collections.unmodifiableMap(dataMapTypes);
     }
 
-    static void takeVanillaSnapshot() {
+    public static void vanillaSnapshot() {
+        LOGGER.debug(REGISTRIES, "Creating vanilla freeze snapshot");
+        takeVanillaSnapshot();
+        LOGGER.debug(REGISTRIES, "Vanilla freeze snapshot created");
+    }
+
+    public static void unfreezeData() {
+        LOGGER.debug(REGISTRIES, "Unfreezing registries");
+        BuiltInRegistries.REGISTRY.stream().filter(r -> r instanceof BaseMappedRegistry).forEach(r -> ((BaseMappedRegistry<?>) r).unfreeze());
+    }
+
+    public static void freezeData() {
+        LOGGER.debug(REGISTRIES, "Freezing registries");
+        BuiltInRegistries.REGISTRY.stream().filter(r -> r instanceof MappedRegistry).forEach(r -> ((MappedRegistry<?>) r).freeze());
+
+        RegistryManager.takeFrozenSnapshot();
+
+        // the id mapping is finalized, no ids actually changed but this is a good place to tell everyone to 'bake' their stuff.
+        fireRemapEvent(ImmutableMap.of(), true);
+
+        LOGGER.debug(REGISTRIES, "All registries frozen");
+    }
+
+    static void fireRemapEvent(final Map<ResourceLocation, Map<ResourceLocation, IdMappingEvent.IdRemapping>> remaps, final boolean isFreezing) {
+        NeoForge.EVENT_BUS.post(new IdMappingEvent(remaps, isFreezing));
+    }
+
+    private static void takeVanillaSnapshot() {
         vanillaSnapshot = takeSnapshot(SnapshotType.FULL);
     }
 
-    static void takeFrozenSnapshot() {
+    private static void takeFrozenSnapshot() {
         frozenSnapshot = takeSnapshot(SnapshotType.SYNC_TO_CLIENT);
     }
 
@@ -282,5 +335,53 @@ public class RegistryManager {
             return false; // Dynamic registry
         }
         return !registry.doesSync();
+    }
+
+    public static void postRegisterEvents() {
+        Set<ResourceLocation> ordered = getRegistrationOrder();
+
+        RuntimeException aggregate = new RuntimeException();
+        for (ResourceLocation rootRegistryName : ordered) {
+            try {
+                ResourceKey<? extends Registry<?>> registryKey = ResourceKey.createRegistryKey(rootRegistryName);
+                Registry<?> registry = Objects.requireNonNull(BuiltInRegistries.REGISTRY.get(rootRegistryName));
+                RegisterEvent registerEvent = new RegisterEvent(registryKey, registry);
+
+                StartupNotificationManager.modLoaderMessage("REGISTERING " + registryKey.location());
+
+                ModLoader.postEventWrapContainerInModOrder(registerEvent);
+            } catch (Throwable t) {
+                aggregate.addSuppressed(t);
+            }
+        }
+        if (aggregate.getSuppressed().length > 0) {
+            LOGGER.error("Failed to register some entries, see suppressed exceptions for details", aggregate);
+            LOGGER.error("Rolling back to VANILLA state");
+            RegistryManager.revertToVanilla();
+            throw aggregate;
+        } else {
+            CommonHooks.modifyAttributes();
+            SpawnPlacements.fireSpawnPlacementEvent();
+            CreativeModeTabRegistry.sortTabs();
+        }
+    }
+
+    /**
+     * Creates a {@link LinkedHashSet} containing the ordered list of registry names in the registration order.
+     * <p>
+     * The order is Attributes, then the remaining vanilla registries in vanilla order, then modded registries in alphabetical order.
+     * <p>
+     * Due to static init issues, this is not necessarily the order that vanilla objects are bootstrapped in.
+     *
+     * @return A {@link LinkedHashSet} containing the registration order.
+     */
+    public static Set<ResourceLocation> getRegistrationOrder() {
+        Set<ResourceLocation> ordered = new LinkedHashSet<>();
+        ordered.add(Registries.ATTRIBUTE.location()); // Vanilla order is incorrect, both Item and MobEffect depend on Attribute at construction time.
+        ordered.add(Registries.DATA_COMPONENT_TYPE.location()); // Vanilla order is incorrect, Item depends on data components at construction time.
+        ordered.add(Registries.ARMOR_MATERIAL.location()); // Vanilla order is incorrect, ArmorItem depends on armor materials at construction time.
+        ordered.addAll(BuiltInRegistries.getVanillaRegistrationOrder());
+        ordered.addAll(BuiltInRegistries.REGISTRY.keySet().stream().sorted(ResourceLocation::compareNamespaced).toList());
+        return ordered;
     }
 }
